instances grain_finished_beef and lean_beef were NOT merged together
Query: in the Beef class, should the instances 'Grain-finished beef' and 'Lean beef' be merged together as semantic and ontologic duplicates? yes or no answer only.
Answer:
	 No.

While both grain-finished beef and lean beef are types of beef, they refer to different characteristics of the meat. Grain-finished beef refers to cattle that have been fed grains during the finishing phase of production, resulting in marbling and flavor. Lean beef, on the other hand, refers to cuts of beef with less fat content. Therefore, it would not be appropriate to merge these two concepts as they describe distinct aspects of the beef product.
meaning that instances grain_finished_beef and lean_beef are different.
Query cache in files:
- /kgfiller/data/cache-9fdbea14aa861cad35bee9364c0f3d9b269215668caa1ae2ec97803e8c4aa420.yml