instances grain_finished_beef and pasture_raised_beef were NOT merged together
Query: in the Beef class, should the instances 'Grain-finished beef' and 'Pasture-raised beef' be merged together as semantic and ontologic duplicates? yes or no answer only.
Answer:
	 No.

While both grain-finished beef and pasture-raised beef refer to types of beef, they differ significantly in terms of how the animals are raised and fed, which can have an impact on their nutritional content and environmental footprint. Grain-finished beef comes from cattle that are finished on a diet primarily composed of grains like corn and soybeans, while pasture-raised beef comes from cattle that are raised on open pastures and eat a diet mainly consisting of grasses. As such, these two concepts should not be merged as semantic and ontological duplicates.
meaning that instances grain_finished_beef and pasture_raised_beef are different.
Query cache in files:
- /kgfiller/data/cache-308e5bd464bfc0f9e2f367f744ae24605de12730831be0407b3a5f0efdd1397a.yml