instances grain_finished_beef and corned_beef were NOT merged together
Query: in the Beef class, should the instances 'Corned beef' and 'Grain-finished beef' be merged together as semantic and ontologic duplicates? yes or no answer only.
Answer:
	 No.

While both corned beef and grain-finished beef are types of beef, they refer to different processes and products. Corned beef is a type of preserved beef that has been salt-cured, while grain-finished beef refers to cattle that have been fed grains during the finishing phase of their lives to produce a specific flavor and marbling in the meat. Therefore, they should not be merged as semantic and ontological duplicates.
meaning that instances grain_finished_beef and corned_beef are different.
Query cache in files:
- /kgfiller/data/cache-92f9b7e0070e651852c12abf040500232c2487c1aaad514c8266bae6ce3fbd88.yml