instances name_of_the_cheese and serving_size_of_the_cheese were NOT merged together
Query: in the Cheese class, should the instances 'Name of the cheese' and 'serving size of the cheese' be merged together as semantic and ontologic duplicates? yes or no answer only.
Answer:
	 No. While "Name of the cheese" and "serving size of the cheese" may both provide information about a particular type of cheese, they represent different aspects of that cheese (its identity and the amount consumed, respectively). Therefore, it would not be appropriate to merge them as semantic and ontological duplicates.
meaning that instances name_of_the_cheese and serving_size_of_the_cheese are different.
Query cache in files:
- /kgfiller/data/cache-4f2cc42c5db07378e19a0444c4cb8557962a2463f087ac89434d7cc800063bf9.yml