instances serving_size_of_the_cheese and shredded_mozzarella_cheese were NOT merged together
Query: in the Cheese class, should the instances 'Shredded mozzarella cheese' and 'serving size of the cheese' be merged together as semantic and ontologic duplicates? yes or no answer only.
Answer:
	 No.

While both "shredded mozzarella cheese" and "serving size of the cheese" are related to cheese, they refer to different concepts and thus should not be merged as semantic and ontological duplicates.

"Shredded mozzarella cheese" refers to a specific type of cheese that has been shredded into small pieces, while "serving size of the cheese" refers to the amount of cheese that is recommended for consumption in one sitting. These two concepts have distinct meanings and uses in nutrition and food preparation contexts.
meaning that instances serving_size_of_the_cheese and shredded_mozzarella_cheese are different.
Query cache in files:
- /kgfiller/data/cache-46ecaa64fec5f75e28fbd8f14364d69c29b20401f685d6ed5964f26df7cc4a1d.yml