instances cholesterol_per_serving and sodium_per_serving were NOT merged together
Query: in the Cheese class, should the instances 'Cholesterol per serving' and 'Sodium per serving' be merged together as semantic and ontologic duplicates? yes or no answer only.
Answer:
	 No.

While both cholesterol and sodium are related to health and nutrition, they are not duplicates of each other from either a semantic or an ontological perspective. Semantically, "cholesterol" and "sodium" have distinct meanings and refer to different nutritional components found in cheese. Ontologically, they belong to separate categories: cholesterol is a type of lipid (fat), while sodium is an electrolyte and mineral. Therefore, it would not be appropriate to merge these two attributes into one field.
meaning that instances cholesterol_per_serving and sodium_per_serving are different.
Query cache in files:
- /kgfiller/data/cache-2a3e456e253a3cd9fe7645e08883874c32c1d5aeef0fb870adebefd16c1289bb.yml