instances carbohydrates_per_serving and vitamin_a_content_per_serving were NOT merged together
Query: in the Cheese class, should the instances 'Carbohydrates per serving' and 'Vitamin A content per serving' be merged together as semantic and ontologic duplicates? yes or no answer only.
Answer:
	 No.

Carbohydrates per serving and Vitamin A content per serving are not semantic or ontological duplicates because they represent different nutritional properties of cheese. Carbohydrates per serving pertains to the amount of carbohydrate present in a given serving size of cheese, while Vitamin A content per serving refers to the quantity of vitamin A found in that same serving size. These two attributes provide distinct information about the nutritional value of cheese and do not overlap conceptually or semantically. Therefore, it would not be appropriate to merge them into a single attribute.
meaning that instances carbohydrates_per_serving and vitamin_a_content_per_serving are different.
Query cache in files:
- /kgfiller/data/cache-fcaffcfebfe4131164a00e146da502f2a136cdb1db242e5a0cc452428f3f4eba.yml